// shared/core/gen/ocean/api/ocean_pb.ts
export enum OceanDeviceStatuses {
  /**
   * @generated from enum value: OCEAN_DEVICE_STATUS_UNSPECIFIED = 0;
   */
  OCEAN_DEVICE_STATUS_UNSPECIFIED = 0,

  /**
   * @generated from enum value: OCEAN_DEVICE_STATUS_BOOTSTRAPPED = 1;
   */
  OCEAN_DEVICE_STATUS_BOOTSTRAPPED = 1,

  /**
   * @generated from enum value: OCEAN_DEVICE_STATUS_PROVISIONING = 2;
   */
  OCEAN_DEVICE_STATUS_PROVISIONING = 2,

  /**
   * @generated from enum value: OCEAN_DEVICE_STATUS_IN_USE = 3;
   */
  OCEAN_DEVICE_STATUS_IN_USE = 3,

  /**
   * @generated from enum value: OCEAN_DEVICE_STATUS_DECOMMISSIONED = 4;
   */
  OCEAN_DEVICE_STATUS_DECOMMISSIONED = 4,
}
// ...
export enum OceanDeviceStates {
  /**
   * @generated from enum value: OCEAN_DEVICE_STATE_UNSPECIFIED = 0;
   */
  OCEAN_DEVICE_STATE_UNSPECIFIED = 0,

  /**
   * @generated from enum value: OCEAN_DEVICE_STATE_EMBARKED = 1;
   */
  OCEAN_DEVICE_STATE_EMBARKED = 1,

  /**
   * @generated from enum value: OCEAN_DEVICE_STATE_DISEMBARKED = 2;
   */
  OCEAN_DEVICE_STATE_DISEMBARKED = 2,
}
// Retrieve enum metadata with: proto3.getEnumType(OceanDeviceStates)
proto3.util.setEnumType(OceanDeviceStates, "ocean.OceanDeviceStates", [
  { no: 0, name: "OCEAN_DEVICE_STATE_UNSPECIFIED" },
  { no: 1, name: "OCEAN_DEVICE_STATE_EMBARKED" },
  { no: 2, name: "OCEAN_DEVICE_STATE_DISEMBARKED" },
]);

/**
 * @generated from enum ocean.OceanCircuitStatuses
 */
export enum OceanCircuitStatuses {
  /**
   * @generated from enum value: OCEAN_CIRCUIT_STATUS_UNSPECIFIED = 0;
   */
  OCEAN_CIRCUIT_STATUS_UNSPECIFIED = 0,

  /**
   * @generated from enum value: OCEAN_CIRCUIT_STATUS_BOOTSTRAPPED = 1;
   */
  OCEAN_CIRCUIT_STATUS_BOOTSTRAPPED = 1,

  /**
   * @generated from enum value: OCEAN_CIRCUIT_STATUS_PROVISIONING = 2;
   */
  OCEAN_CIRCUIT_STATUS_PROVISIONING = 2,

  /**
   * @generated from enum value: OCEAN_CIRCUIT_STATUS_IN_USE = 3;
   */
  OCEAN_CIRCUIT_STATUS_IN_USE = 3,

  /**
   * @generated from enum value: OCEAN_CIRCUIT_STATUS_DECOMMISSIONED = 4;
   */
  OCEAN_CIRCUIT_STATUS_DECOMMISSIONED = 4,
}
// ...
export enum OceanCircuitStates {
  /**
   * @generated from enum value: OCEAN_CIRCUIT_STATE_UNSPECIFIED = 0;
   */
  OCEAN_CIRCUIT_STATE_UNSPECIFIED = 0,

  /**
   * @generated from enum value: OCEAN_CIRCUIT_STATE_EMBARKED = 1;
   */
  OCEAN_CIRCUIT_STATE_EMBARKED = 1,

  /**
   * @generated from enum value: OCEAN_CIRCUIT_STATE_DISEMBARKED = 2;
   */
  OCEAN_CIRCUIT_STATE_DISEMBARKED = 2,
}