// shared/core/gen/harbor/api/harbor_pb.ts
export enum HarborJobTypes {
  /**
   * @generated from enum value: HARBOR_JOB_TYPE_UNSPECIFIED = 0;
   */
  HARBOR_JOB_TYPE_UNSPECIFIED = 0,

  /**
   * @generated from enum value: HARBOR_JOB_TYPE_EMBARK = 1;
   */
  HARBOR_JOB_TYPE_EMBARK = 1,

  /**
   * @generated from enum value: HARBOR_JOB_TYPE_DISEMBARK = 2;
   */
  HARBOR_JOB_TYPE_DISEMBARK = 2,
}
// Retrieve enum metadata with: proto3.getEnumType(HarborJobTypes)
proto3.util.setEnumType(HarborJobTypes, "harbor.HarborJobTypes", [
  { no: 0, name: "HARBOR_JOB_TYPE_UNSPECIFIED" },
  { no: 1, name: "HARBOR_JOB_TYPE_EMBARK" },
  { no: 2, name: "HARBOR_JOB_TYPE_DISEMBARK" },
]);

/**
 * @generated from enum harbor.HarborJobStatuses
 */
export enum HarborJobStatuses {
  /**
   * @generated from enum value: HARBOR_JOB_STATUS_UNSPECIFIED = 0;
   */
  HARBOR_JOB_STATUS_UNSPECIFIED = 0,

  /**
   * @generated from enum value: HARBOR_JOB_STATUS_INIT = 1;
   */
  HARBOR_JOB_STATUS_INIT = 1,

  /**
   * @generated from enum value: HARBOR_JOB_STATUS_RUNNING = 2;
   */
  HARBOR_JOB_STATUS_RUNNING = 2,

  /**
   * @generated from enum value: HARBOR_JOB_STATUS_FAILED = 3;
   */
  HARBOR_JOB_STATUS_FAILED = 3,

  /**
   * @generated from enum value: HARBOR_JOB_STATUS_DONE = 4;
   */
  HARBOR_JOB_STATUS_DONE = 4,
}
// ...
export enum HarborStepStatuses {
  /**
   * @generated from enum value: HARBOR_STEP_STATUS_UNSPECIFIED = 0;
   */
  HARBOR_STEP_STATUS_UNSPECIFIED = 0,

  /**
   * @generated from enum value: HARBOR_STEP_STATUS_INIT = 1;
   */
  HARBOR_STEP_STATUS_INIT = 1,

  /**
   * @generated from enum value: HARBOR_STEP_STATUS_RUNNING = 2;
   */
  HARBOR_STEP_STATUS_RUNNING = 2,

  /**
   * @generated from enum value: HARBOR_STEP_STATUS_FAILED = 3;
   */
  HARBOR_STEP_STATUS_FAILED = 3,

  /**
   * @generated from enum value: HARBOR_STEP_STATUS_COMPLETED = 4;
   */
  HARBOR_STEP_STATUS_COMPLETED = 4,
}